bluetooth: Fix a crash during device removal and remove unnecessary checks
This fixes a improper access to the device which has been removed during
the processing of RemoveDevice.

This removes the redundant checks when setting the values of device
properties, since the value guard is guaranteed by the Property Type.
And this also removes the unnecessary traversal of discovered device.

BUG=None
TEST=Run unit tests and perform manual tests with a LE keyboard with the
following steps.
(0) Start discovery and connect to the keyboard.
(1) Confirm it's connected and then disconnect it.
(2) Confirm it's disconnected and then re-connect it.
(3) Remove it and confirm it's removed without crashing.

Change-Id: Ica8acf5fec18176c1f74ca817f3650d788dc8ab6
Reviewed-on: https://chromium-review.googlesource.com/1615593
Commit-Ready: Miao-chen Chou <mcchou@chromium.org>
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>