metrics: reparent orphans and tolerate chrome children
This fixes one observed and one possible bug.

The observed bug is that Chrome execs a variety of
executables and the current checks were too strict.

The other change is from not wanting rely on specific
semantics of /proc.  Specifically, we conservatively
assume that our snapshot may capture a process whose
parent has died, but the child has not been reparented
yet. Conservatively we reparent that process to init.
This should be rare (if it happens at all) so that the
workaround will not affect stats, but without the
workaround the resulting crash could be annoying.

BUG=chromium:963659
TEST=unit test, also ran on device

Change-Id: Ida69dba066c03c5647a4c474f73b16e02d322c44
Reviewed-on: https://chromium-review.googlesource.com/1614710
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>