shill: Remove assumption that RpcIdentifier is a std::string
Shill currently uses RpcIdentifier and std::string interchangeably in
the context of RPC identifiers. Semantically, however, an
RpcIdentifier is a valid DBus object path, not just any string. In
particular, KeyValueStore serializes RpcIdentifiers and strings
differently. In order to create a templated KeyValueStore
implementation with no special-cases in the interface, it will be
necessary for RpcIdentifier to be a distinct type from string.

This change replaces uses of string with RpcIdentifier when the
semantics of the data are of an RPC identifier. A later change
will make RpcIdentifier a typedef of dbus::ObjectPath, as that is what
is meant semantically when RpcIdentifiers are used.

BUG=chromium:936111
TEST=All unit tests are passing.

Change-Id: Ib2eef061ed38c1399e3951d7a0e64e6f45ca8dd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1646936
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>