shill: Make RpcIdentifier a dbus::ObjectPath
Semantically, the only difference between a std::string and an
RpcIdentifier is that an RpcIdentifier is assumed to be something that
one could get from dbus::ObjectPath::value(). Since RpcIdentifier must
be a different type from std::string in order for templatization of
KeyValueStore and PropertyStore to work cleanly, making it an alias to
dbus::ObjectPath seems like a logical choice.

BUG=chromium:936111
TEST=All unit tests are passing.

Change-Id: I77778f0d912bc6a247330bc259f152205da2de55
Reviewed-on: https://chromium-review.googlesource.com/1646937
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>