power: Refactor: Remove stale dark resume code.
Previously (legacy dark resume) if wakes are triggered by non-input
devices, we used to resume only specific drivers. So on user input when
in dark resume, we have to do a suspend/resume cycle to the driver
level to exit out of dark resume. This can fail sometimes. When it fails
we used to simply shutdown. This is not the case anymore. So we do not
need this complicated code.

BUG=chromium:964510
TEST=test cases pass.

Change-Id: I49097d2a7925a679bf25f489800bc373564d6e3a
Reviewed-on: https://chromium-review.googlesource.com/1622752
Commit-Ready: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>