smbprovider: Save and restore password to the daemon-store.
The password file is stored in the user's daemon-store directory, which
is part of the user's cryptohome and encrypted at rest, and only
accessible by the smbproviderd user. Because Mount() requests come from
the logged-in profile, it is safe to assume the daemon-store directory
exists.

If the user explicitly removes the share, the password file should be deleted.

BUG=chromium:891462
TEST=Run unit tests
TEST=Deployed with dependent Chrome change and observe password restored after
     logout/login

Change-Id: Id6e218a7a86cbddb6cb531123fcc16f413fa1f0a
Reviewed-on: https://chromium-review.googlesource.com/1663745
Tested-by: Anand Mistry <amistry@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>