hardware_verifier: Change text format from prototxt to JSON
In proto3, a field with default value is equivalent to an unsetted
field.  It turns out if a value is `false` or the first enum item,
the field will not be shown in prototxt format, which is confusing.
This CL changes the format to JSON since the converter accepts
an option to always print all fields.

BUG=b:130780188
TEST=FEATURES=test emerge-sarien hardware_verifier; manual test

Change-Id: I8fe78e710ec0c122a130e13588ee120bcaf40e06
Reviewed-on: https://chromium-review.googlesource.com/1627398
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Yong Hong <yhong@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hong-Min Chu <hmchu@chromium.org>