common-mk: make proto_library accept all_dependent_pkg_deps
This change is for gdisp package to compile after migration.
https://chrome-internal-review.googlesource.com/c/chromeos/platform/gdisp/+/1330567

Though it is a bit strange to include all_dependent_pkg_deps inside
proto_library (as there is no inherent reason to do so), one-to-one
translation of the package from GYP is to do so. As the package is
large, I'd like to not do anything more than translating it to GN, hence this
change.

TEST=new pakcage using all_dependent_pkg_deps inside proto_library.
BUG=chromium:953642
BUG=chromium:923986

Change-Id: I664ffb0a2ec2828e2c4429fec8b5b8149df70151
Reviewed-on: https://chromium-review.googlesource.com/1628510
Commit-Ready: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>