Kerberos: Add option to remember password
AcquireKerberosTgt() can now be called with a remember_password flag. If
set, the password is stored on disk and loaded if AcquireKerberosTgt()
is called with an empty password. Chrome should only call this if it
knows that the daemon has a stored password. For this reason,
ListAccounts() reports this in the has_remembered_password field.

If AcquireKerberosTgt() is called without the remember_password or if
the account is removed, the stored password is cleared.

BUG=chromium:952243
TEST=Unit tests

Change-Id: I9b0519adbe990ca60372f2d0837796e33c4943cd
Reviewed-on: https://chromium-review.googlesource.com/1631690
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>