power: Implement shutdown after x.
This codes sets an alarm for kShutdownAfterSecsPref on path to suspend
(if one is not already running).
On a wake up
  If in full resume :
    stops any running alarm.

  If in dark resume and not on line power:
    Forces a device shutdown instead of re-suspend if the alarm has
    already fired.
    If the alarm has not fired simply slips back to suspend.

  If in dark resume and on line power, simply slips back to suspend.
  Note that, a new alarm will be set again for kShutdownAfterSecsPref if
  one is not running.

On a transition to full resume from dark resume
  Stops the running alarm.

Note that this change does not handle shutting down the device if the
battery is below a configured threshold. This is to follow in a seperate
cl.

Test cases are in the cl that this depends on.

BUG=chromium:964510
TEST=deploy on device and make sure it works.

Change-Id: Ic6ebafcb4bb044c7d0a4f46c82bd8b22a7694abe
Reviewed-on: https://chromium-review.googlesource.com/1627056
Commit-Ready: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>