system_api: remove STORAGE_CRYPTOHOME_DOWNLOADS from concierge API
User's Downloads directory is not accessible to concierge/crosvm (crosvm
user does not have access to chronos' files), so there is no reason of
having this option as storage location for images.

The plan is to pass an image's file descriptor if access outside of
/home/root/<cryptohome_id>/{crosvm|pvm} is needed.

This removes STORAGE_CRYPTOHOME_DOWNLOADS and reorders the rest so
STORAGE_CRYPTOHOME_PLUGINVM moves over by one. It is OK as Chrome is not
using STORAGE_CRYPTOHOME_PLUGINVM yet.

BUG=b:130304430
TEST=Build and boot

Change-Id: Ie398010a2522333ab12900e29e87e94d4042b260
Reviewed-on: https://chromium-review.googlesource.com/1626887
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>