hardware_verifier: Output the whitelisted generic device info
In addition to output the the qualification status of the components
which are recorded in the verification spec, this CL lets the
verification report include the generic component values probed
by the runtime probe.  To prevent the service from leaking any
privacy data, this program will output the fields that are listed
in the whitelist in the verification spec.

BUG=chromium:926825, b:130780188, b:133485789
TEST=manual test

Change-Id: I9f39e014d50ac0b21fe8fff579c6c1ab374f4973
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1630332
Reviewed-by: Hong-Min Chu <hmchu@chromium.org>
Tested-by: Yong Hong <yhong@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Commit-Queue: Chun-ta Lin (ping on chat if not responsive) <itspeter@chromium.org>