arc: networkd: Generalize ARC lifecycle notifications.
This patch removes the signalling mechanism used to indicate when ARC++
starts and stops and replaces it with a simple text protocol over an
abstract domain socket. This allows us to use the same upstart scripts
without overlay for detecting both ARC++ and ARCVM. Recall this is
intended to be temporary until the guests connect directly to the
daemon.

This change in turn requires message propagation to the child tasks
(both adb proxy and ip helper) and the appropriate refactoring was
done. The expansion of the protos is required anyway eventually to
support notifications for multiple guests.

BUG=b:133706953
TEST=unit tests, tasts, some autotests
TEST=manual testing with arc++ and arcvm, verified play store and adb
TEST=suite:bvt-arc, CtsHostsideNetworkTests
Cq-Depend: chromium:1648549
Change-Id: Ia6c79d66b40cb63a6269ef9271089398b8724ae1
Reviewed-on: https://chromium-review.googlesource.com/1631854
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Ready: Garrick Evans <garrick@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>