arc: networkd: SO_REUSEADDR needed for adb-proxy on localhost.
This patch sets the SO_REUSEADDR on the listening socket so it can
be reused. If adb is connected through the proxy and the user logs out
then when they log back in, adb-proxy will not be able to rebind and
adb is unusable until the server is killed and ARC++ is cycled again.

BUG=None
TEST=verified manually

Change-Id: Icf99df19e958ec404180af0f517d746d11499de3
Reviewed-on: https://chromium-review.googlesource.com/1634889
Commit-Ready: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>