crash: Remove last env var and fix command line testing
This removes the OVERRIDE_PAUSE_SENDING env var and replaces it with a
command line flag. It also fixes some of the tests for parsing the
command line as this change brought out the existing flaw in that, so it
needed to be fixed as well.

BUG=chromium:391887, chromium:958968
TEST=Unit and autotests pass

Cq-Depend: chromium:1635798
Change-Id: I7e24bbc6d29f51c81970666f8a0df1f5dee4c559
Reviewed-on: https://chromium-review.googlesource.com/1635833
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Tim Zheng <timzheng@chromium.org>