biod: Change CrosFpFirmware to allow mocking
We allow a CrosFpFirmware to remain in the default
kUninitialized state if no file path is given in the
constructor. This mimics the base::File behavior.

This is needed because the mock class needs to be
able to inject fake versions for a CrosFpFirmware,
without presenting an actual firmware file.

The issue with the current implementation is that it
attempts to decode the firmware file in the single
constructor. The functionality of a superclass
constructor cannot be overridden (even with a virtual
function).

BUG=b:117182903
TEST=FEATURES=test emerge-nocturne biod
TEST=FEATURES=test emerge-nami biod

Change-Id: I74f7fce3aaa3cfae10da2d826be92f674c407980
Reviewed-on: https://chromium-review.googlesource.com/1636173
Tested-by: Craig Hesling <hesling@chromium.org>
Commit-Ready: Craig Hesling <hesling@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>