power: Set the default keyboard wakeup capability to enabled.
As part of crrev.com/c/1069220, we deleted the udev rule that
tries to control the wakeup capability of internal keyboard. The reason
for the change was powerd and EC seem to get into a race to control
wakes from internal keyboard. Please look at crbug.com/836279 for more
details. After this change powerd stopped worrying abount internal
keyboard.
This will be a problem if the default value of power/wakeup
property of the internal keyboard is set to disabled for 2 reasons.
  1. If the default state is disabled, and if the user presses a key
     when the kernel is in the suspend path, power core in the
     kernel will not abort the suspend.
  2. Dark resume will not work as we cannot attribute wakeups to
     keybaord since the wakeup count will not increment if the
     property is set to disabled.

To fix this on the ARM side, we set the default state to
enabled(crrev.com/c/1101959). On the intel side we use atkbd driver,
which is a generic driver. I highly doubt a patch to set the default
state will be accepted. Thus instead set the default value to enabled
in powerd.

How was this not a problem on recent devices?
It so happened that most of the devices that we build recently have side
buttons. Thus we install a option udev rule
93-powerd-tags-keyboard-side-buttons.rules which enables wakeup.

How was this not a problem on nocturne?
Though nocturne does not have side buttons, nocturne does not use atkbd
driver for internal keyboard. Since the internal keyboard is on usb bus,
wake is enabled by default.

Will this change cause any regressions on older devices?
It should not. First on intel devices, only kernel honours this
property. ACPI does not. Thus if keyboard causes any spurious wakes,
previously device would end up in a suspend resume loop as kernel
suspends and ec would wake the system immediately. Now the suspend
fails. Thus the system would try for 10 times and shuts down instead.
Thus this should only improve the situation.
On ARM devices this property is set to enabled by default anyway.

BUG=b:133262012
TEST=Run power_WakeSources and make sure the INTERNAL_KEYBOARD passes.

Change-Id: I6f05c7bd01cc542b912fa20bc94b2634efce0dc3
Reviewed-on: https://chromium-review.googlesource.com/1640180
Commit-Ready: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>