vm_tools: Pass name of stateful device to maitred.
Stateful device name used to be hardcoded in the maitred. When we
introduce virtio-pmem and start using it for VM's rootfs, the naming of
devices changes so there need to be a way to pass the name of stateful
device from concierge to maitred.

BUG=None
TEST=vm.CrostiniStartEverything

Change-Id: If938b51694a909038182d1ff5420f20ed90e042c
Reviewed-on: https://chromium-review.googlesource.com/1638117
Commit-Ready: Jakub Staroń <jstaron@google.com>
Tested-by: Jakub Staroń <jstaron@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>