bluetooth: Fix lingering connection attempt and session during device removal
This includes to following fixes.
- Reply to D-Bus calls to connect/disconnect/pair to prevent D-Bus timeout
upon RemoveDevice().
- Disconnect the device upon device removal.
- Clean up |connect_sessions_| for the connect/disconnect calls initiated
by the stack upon RemoveDevice().
- Add DisconnectInternal to handle the disconnection and reply to error.
- Constructing the ConnectSession inside ConnectInternal() and
DisconnectInternal() to make the tracking of ConnectSession less
error-prone.

This considers the following scenarios during device removal:
a) There is an ongoing connection attempt with the device. The attempt was
initiated by either the stack or D-Bus client.
b) There is active connection with the device.
c) There is an ongoing disconnection attempt with the device. The attempt
was initiated by either the stack or D-Bus client.
d) The device is discovered nor paired or connected.
e) There is an ongoing pairing attempt with the device. The attempt was
initiated by D-Bus client.
f) The device is paired but not connected.

BUG=chromium:863613
TEST=(1) All unit tests pass.
     (2) On Bluetooth setting page, discover the target device and
     perform device removal with (a)(d)(f) scenarios.
     (3) In bluetoothctl console, perform device removal with
     (a)(c)(d)(e) scenarios.

Change-Id: I83bd32b8f441022db1404351f9183319d54c2a75
Reviewed-on: https://chromium-review.googlesource.com/1647314
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>