arc: net: Allow multicast forwarder to start with layer 3 information
This patch changes arc-networkd multicast forwarders so that they can be
instantiated and bound to a device without layer 3 knowledge. When the
IPv4 address of the outbound physical interface is needed during
forwarding, it is always queried from the interface and never cached.

This allows the forwarders to work when the network connection on top
of a particular network interface is not yet ready when the forwarders
are started.

This also allows the forwarders to keep working when IPv4 provisioning
changes.

To support joining a multicast group on an interface without any IP
address bound, the newer struct ip_mreqn is used in conjunction with the
interface index for the setsockopt IP_ADD_MEMBERSHIP system call.

BUG=b:112661642
BUG=b:134713711
TEST=1) multinet off, eth0 and wlan0 connected: logged in, checked that
switching the default network in shill triggers the multicast forwarders
to switch. Packets are forwarded correctly.
     2) multinet off, only wlan0 connected: logged in and out while also
disabling and enabling WiFi. Checked that the multicast forwarders
always recover once layer 3 is provisioned on WiFi.
     3) multinet on, eth0 and wlan0 connected: same tests as 2).

Change-Id: I61f3f88f29d490a0cd159fc5cac10c4b0b9fe52f
Reviewed-on: https://chromium-review.googlesource.com/1648542
Commit-Ready: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>