crash: lock |core_pattern| once crash-reporter is initialized
For kernels older that linux-4.4, lock_core_pattern was added to
prevent modifications to the core_pattern once crash-repoter
is done with initialization. This patch also adds code
to handle failure to reset the core_pattern if it is locked:
this can occur during "crash-reporter --clean_shutdown"
on devices that support the core pattern lock.

The locking of the core_pattern is conditional on the
image type (disabled on dev images) to allow for a
regression test in logging_UserCrash.

BUG=chromium:960445
TEST=logging_UserCrash passes.

Cq-Depend: chromium:1658734
Change-Id: I59da5dd86339a3457bbd4729c9eea683abee4d7f
Reviewed-on: https://chromium-review.googlesource.com/1652001
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>