arc: networkd: Handle race condition when setting IPv6 configuration.
This is a pre-emptive patch to allow for using netlink events to drive
device enablement. It is necessary because once this happens, IPv6
configuration will be attempted as soon as the device is running which
is independent of the container state, in fact, the user may not even
be logged in. Of course this prevents the assignment from succeeding.

So this patch uses the same mechanism as for enabling inbound traffic
for the legacy device by gating on the container being ready. It also
accounts for the possibility that Android has not (yet) setup the
routing table and will retry if necessary. Finally, this patch also
changes many CHECKs to LOG(DFATAL) to avoid crashing the process.

BUG=b:133194026
TEST=some tasts and autotests, lots of manual testing on nocturne

Change-Id: Id8688d09e66e2ab5106ce555ecc098c57ee950ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1654754
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>