debugd: Remove dead code for AnonymizeLogMap
Was looking at this file and noticed this wasn't being used so cleaning
it up. This specific call wasn't used for anonymization, we have the
anonymizer as a member variable and that's being used at other call
sites in this class to do anonymization.

BUG=None
TEST=Unit tests pass

Change-Id: I454686f7af7758b4296e5bb81277fc6894edbda5
Reviewed-on: https://chromium-review.googlesource.com/1659428
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>