crash: Don't send crashes if the network is offline
This uses a D-Bus call to shill to check the ConnectionState of the
network and will not send crashes if we get back a valid value and it is
not "online".  Build file was slightly reorganized so that only test
source is in the test target now and renamed the existing proxy_ var to
avoid confusion.

BUG=chromium:200191
TEST=Manually verfied, unit tests pass
Cq-Depend: chromium:1664982
Change-Id: If98ce72463bcfc3212ffdd6523176554b23385a9
Reviewed-on: https://chromium-review.googlesource.com/1661192
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>