modemfwd: record carrier for main firmware journal entries
If the carrier ID is necessary to decide which main firmware we
should be flashing, then we have to record both the device and
carrier for each entry. Instead of inferring whether or not the
journal entry is for main or carrier firmware based on the
presence or absence of the carrier ID in the journal entry, we
now keep an explicit entry type.

The RestartOperation function will still be able to read V1
journal entries, since a journal entry may persist across an
update.

BUG=b:135009654
TEST=unit tests

Change-Id: I5818a03fe4d702de1947eba90367bd2770faea41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1663251
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Auto-Submit: Eric Caruso <ejcaruso@chromium.org>