arc: net: Increase IPv6 provisioning delay by 2 seconds
As commented in router_finder, sendto() on an IPPROTO_ICMP6 socket can
return EADDRNOTAVAIL if done too quickly after preparing the socket. As
of now the reason is still unknown as EADDRNOTAVAIL is not a documented
error case of sendto(). Investigations in the linux kernel sources for
the socket protocol handler of IPPROTO_ICMP6 did not revealed anything.
One possibility which was not explored was a possible bug in the third
party ndp library.

The failure of sendto() on EADDRNOTAVAIL can prevent IPv6 provisioning
for ARC, which is a user facing issue and a source of spurious errors
for CTS verifications. This patch adds 2 seconds to the provisioning
delay in order to reduce the failure rate.

The impact is limited to IPv6 only networks since on IPv4 only networks
or dual stack networks, connectivity over IPv4 is available typically
almost immediately thanks to the host network already being provisioned.

BUG=b:135545546
TEST=Without the additional delay, IPv6 provisioning on login can fail
up to 20~25% of the time on eve with ARC multinetworking on (over ~20
trials). With the additional delay, no failure were observed over 10
trials.

Change-Id: I0c086a54a759cd6de23d6b4aa28a0fa3745fbf65
Reviewed-on: https://chromium-review.googlesource.com/1669384
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>