system_api: Add a new GetStatus dbus method and signal for update_engine
based on previous CL:1127279

This new API is added to replace the old GetStatus API because GetStatus
is not easily extendable. We use this added protobuf to communicate with
othe processes regarding the status of the update_engine.

Basically we are adding a new DBus method GetStatusAdvanced and a new
DBus signal StatusUpdateAdvanced.

BUG=chromium:977320
TEST=precq, unittests

Change-Id: Ia0fc7c8cc0af091db9a5d9955816c86e7f8117e6
Reviewed-on: https://chromium-review.googlesource.com/1670444
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>