shill: routing_table: Replace per-table route on receipt of main table route
Previously, routes added to the main routing table would simply be
removed if shill is aware of a corresponding entry existing in the
per-device table. This change modifies such behavior to have these
entries replaced instead. This prevents a potential scenario in which
a default route fails to be added to a per-device table but shill
believes it to be there, leading to the inability for future route
additions to actually make their way from the main routing table to
the per device table.

This also allows for the 'expires' property to be used by routes in
per-device tables in the future, as periodic routes added by the
kernel would be able to update the expiration time.

BUG=chromium:941597
TEST=-All unit tests are passing.
     -v6 RA entries getting updated upon receipt of new RA (tested by
      adding an expire value to existing default routes in per-device
      tables, which shill itself doesn't do).

Change-Id: I52d4b8de6ace4a36e02822537ef7eb7cf7b7d936
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1672252
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Auto-Submit: Alex Khouderchah <akhouderchah@chromium.org>