shill: routing_table_entry: Rm unnecessary operator=s and {con,de}structors
Routing{Table,Policy}Entry explicitly defined assignment operators
that are equivalent to the default ones. Not only does this bloat the
interface and make it more cumbersome to add new fields, but it also
prevents the generation of the move operator and
constructor. Similarly, the default destructors and copy constructor
are unnecessary and are removed.

BUG=chromium:930280
TEST=All unit tests are passing.

Change-Id: I436cd7567f00989dcd45dfe34de863e056656ba3
Reviewed-on: https://chromium-review.googlesource.com/1672248
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>