dlcservice: Use DBus protobuf capabilities for update_engine DBus methods
DBus have the capability to pass protobuf message directly in the
methods and signals without the need for serializing it manually. This
CL uses those types of arguments which is almost used everywhere in the
platform2 code base. This allows users of the API to not care about
serializing/deserializing the protobuf and their implementation will be
much cleaner.

BUG=chromium:978672
TEST=unittest, precq

Cq-Depend: chromium:1672021
Change-Id: I699b00bb84ec7f4ac384ae490aca91c7f5069f23
Reviewed-on: https://chromium-review.googlesource.com/1672684
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>