login: Add wallpaper switch for unibuild devices with a custom wallpaper
In order to add the custom wallpaper in the "My Images" tab of the
wallpaper app, the "default-wallpaper-is-oem" switch must be in the
Chrome flags. Adding this for the case of OEM wallpapers using
cros-config (unibuild devices).

BUG=b:134987263
TEST=ran on nautilus and verified that default wallpaper appears
in the "My Images" section

Change-Id: I78298e29f335dd364f5ccd233127f7687a97e1b8
Reviewed-on: https://chromium-review.googlesource.com/1671607
Tested-by: May Lippert <maybelle@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>