chromeos-config: handle case of model name equals to wl_sig_id.
Normally, in whitelabel project the key-id will not equal to model
name because key-id is based on LOEM's name. But in case of leveraging
existed model to the new one with only marketing name change,
whitelabel_tag is used. As a result, we have the chance of key-id equal
to model name.

The solution here is to filter this situation and not mark this
firmware info of config to have_image=False.

BUG=b:127342204
TEST=`pack_firmware.py` and verify the result of Octopus.

Change-Id: I8f3b0252b0228540305ba31b48e9314586f23af4
Reviewed-on: https://chromium-review.googlesource.com/1675635
Tested-by: Marco Chen <marcochen@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>