cryptohome: Skip useless migrations of chall-resp keysets
Avoid trying to migrate the challenge-response keys when it's
unnecessary (currently - always).

The old code was seeing two facts - the TPM is available and that
the keyset is not encrypted via the TPM - and therefore was trying
to perform the migration (re-encryption) of the keyset. This was
anyway producing the same keyset, so this all was just a waste of
resources.

BUG=chromium:842791
TEST=manual

Change-Id: I94f6a68463979eac641767bc34ede74d70997e81
Reviewed-on: https://chromium-review.googlesource.com/1678660
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>