shill: Add routing rules to use non-VPN PPP interface when src address matches
Currently, Connection only creates rules sending traffic whose source
address matches one of the interface's addresses to the corresponding
per-device table when the Device is a physical technology. PPPoE is an
exception to that rule, where a PPP interface should have that same
behavior, despite being virtual. This change allows for this without
compromising the routing logic for PPP interfaces corresponding to
VPN connections (where this new behavior is not desired).

The same argument applies for the PPPDevice created by cellular for
cellular dongles.

BUG=chromium:941597,b:135813021
TEST=-All unit tests are passing.
     -`test_that $DUT network_VpnConnect.l2tpipsec_psk` passes.
     -`run_lab_test.py --board=whirlwind jetstream_Pppoe` passes.

Change-Id: I571d754748ed634c6edc6b911b63a2992fb6db4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1679043
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>