sealed_storage: check policy and size during unseal
This CL stores information about the policy used when sealing and the
size of the original data in the sealed data blob. And verifies that
those parameters are correct when unsealing.

To do that the CL introduces V2 of the storage format for sealed data
and maintains support for legacy V1 when unsealing.

The CL also updates the README to cover the new procedure.

BUG=b:134734017
TEST=1) Unit tests. New unit tests that failed before, and pass now,
        were added: WrongRestoredZPointGarbage, WrongPolicyi and more.
        New unit tests for V1 backwards compatibility were added.
     2) Live tests from CL:1562022 and CL:1603282 (and more).

Change-Id: If67f89454f26a3a4ef58df26a41a93efc8d1d7c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1680072
Tested-by: Andrey Pronin <apronin@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Andrey Pronin <apronin@chromium.org>
Auto-Submit: Andrey Pronin <apronin@chromium.org>