kerberos: Add TgtRenewalScheduler class
TgtRenewalScheduler is responsible for scheduling automatic Kerberos
ticket-granting-ticket (TGT) renewals. Usually, TGTs are valid for 8-10
hours and can be renewed periodically for a week or 2, which refreshes
the validity lifetime each time until the renewal lifetime limit is hit.

TgtRenewalScheduler schedules renewal tasks. If the ticket lifetime is
10 hours, it schedules automatic renewals after 6h0m, 8h24m, 9h21m,
9h44m, 9h53m and 9h 57m. If all these attempts failed, it will send a
notification that the ticket is about to expire. If the ticket was
already expired, e.g. after waking up from a suspended state, it sends
out a notification that the ticket has already expired.

The class is not used yet. This is done in a follow-up CL.

BUG=chromium:952242
TEST=Unit test

Change-Id: Ice4ceac8432b6055098e2773ac4baa11632e3bcf
Reviewed-on: https://chromium-review.googlesource.com/1680417
Tested-by: Lutz Justen <ljusten@chromium.org>
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>