dlcservice: add example programs to be referred to by developer doc
We add two example programs to demo the usage of Install and Uninstall
API respectively. We separate them in two separate programs so to not
confuse developers the relationships of these two APIs. Also it's easier
to refer them to individually in the doc.

We also compile them in BUILD.gn (but not install) so
we make sure they're up-to-date. These two programs will be linked to a
DLC developer doc that is currently WIP.

BUG=chromium:955176
TEST=emerge-kevin-arcnext dlcservice

Change-Id: Idc734a9f1e81a8ef5a873eff8008dfb196d4c69b
Reviewed-on: https://chromium-review.googlesource.com/1684355
Tested-by: Xiaochu Liu <xiaochu@chromium.org>
Commit-Ready: Xiaochu Liu <xiaochu@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Xiaochu Liu <xiaochu@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>